Remove RMQ, Add API POST request method.
