+++ Pipfile
@@ -4,15 +4,12 @@ verify_ssl = true
 name = "pypi"
 
 [packages]
-pika = "*"
 docker = "*"
+flask = "*"
+gevent = "*"
+gunicorn = "*"
 
 [dev-packages]
-flask = "*"
-flask-sockets = "*"
-gevent = "==1.2.2"
-gevent-websocket = "*"
-gunicorn = "*"
 "flake8" = "*"
 pytest = "*"
 pytest-cov = "*"
@@ -25,18 +22,10 @@ python_version = "3.6"
 lint = "flake8"
 test = "py.test tests --cov . --cov-report term-missing -v"
 report = "py.test tests --cov . --cov-report=html"
-
-snekbox = "python snekbox.py"
-snekweb = "gunicorn -w 2 -b 0.0.0.0:5000 --log-level debug -k geventwebsocket.gunicorn.workers.GeventWebSocketWorker snekweb:app"
-
+snekbox = "gunicorn -w 2 -b 0.0.0.0:8060 snekbox:app"
 buildbox = "docker build -t pythondiscord/snekbox:latest -f docker/Dockerfile ."
 pushbox = "docker push pythondiscord/snekbox:latest"
-
 buildboxbase = "docker build -t pythondiscord/snekbox-base:latest -f docker/base.Dockerfile ."
 pushboxbase = "docker push pythondiscord/snekbox-base:latest"
-
 buildci = "docker build -t pythondiscord/snekbox-ci:latest -f docker/ci.Dockerfile ."
 pushci = "docker push pythondiscord/snekbox-ci:latest"
-
-buildweb = "docker build -t pythondiscord/snekboxweb:latest -f docker/Dockerfile.webapp ."
-pushweb = "docker push pythondiscord/snekboxweb:latest"
