+++ docker/Dockerfile
@@ -1,5 +1,7 @@
 FROM pythondiscord/snekbox-base:latest
 
+RUN apk add --update tini
+
 RUN mkdir -p /snekbox
 COPY . /snekbox
 WORKDIR /snekbox
@@ -7,4 +9,5 @@ WORKDIR /snekbox
 RUN pipenv --rm
 RUN pipenv sync
 
+ENTRYPOINT ["/sbin/tini", "--"]
 CMD ["pipenv", "run", "snekbox"]
