+++ templates/index.html
@@ -1,106 +1,14 @@
 <!DOCTYPE html>
-<meta charset="utf-8" />
-<title>snekboxweb</title>
-<script language="javascript" type="text/javascript">
-
-
-let _ready = false
-let snekbox_id
-var output;
-
-snekbox_id = sessionStorage.getItem("snekbox_id");
-console.log(snekbox_id)
-if (snekbox_id == null) {
-    snekbox_id = generate_id()
-    sessionStorage.setItem("snekbox_id", snekbox_id)
-    console.log(snekbox_id)
-}
-
-function init(){
-    output = document.getElementById("output");
-    websocketHandler();
-}
-
-function websocketHandler(){
-    var here = window.location.host;
-    var wsUri = `ws://${here}/ws/`+snekbox_id;
-    websocket = new WebSocket(wsUri);
-    websocket.onopen = function(evt) { onOpen(evt) };
-    websocket.onclose = function(evt) { onClose(evt) };
-    websocket.onmessage = function(evt) { onMessage(evt) };
-    websocket.onerror = function(evt) { onError(evt) };
-}
-
-function onOpen(evt){
-    _ready = true
-    console.log("CONNECTED");
-}
-
-function onClose(evt){
-    _ready = false
-    console.log("DISCONNECTED");
-}
-
-function onMessage(evt){
-    writeToScreen('<span style="color: blue;">RESPONSE: ' + evt.data+'</span>');
-}
-
-function exit(){
-    websocket.close();
-}
-
-function onError(evt){
-    _ready = false
-    writeToScreen('<span style="color: red;">ERROR:</span> ' + evt.data);
-}
-
-function sendMessage(msg){
-    waitForSocketConnection(function(){
-        websocket.send(msg);
-    });
-    console.log("sent message "+msg)
-}
-
-function waitForSocketConnection(callback){
-    setTimeout(
-        function () {
-            if (_ready === true) {
-                if(callback != null){
-                    callback();}
-                return;
-            }
-            else {
-                waitForSocketConnection(callback);}
-
-        }, 500); // milliseconds
-}
-
-function writeToScreen(message){
-    var pre = document.createElement("p");
-    pre.style.wordWrap = "break-word";
-    pre.innerHTML = message;
-    output.appendChild(pre);
-}
-
-function sendFromInput(){
-    var msg = document.getElementById("field1").value;
-    sendMessage(msg)
-}
-
-function generate_id(){
-    return Math.random().toString(36).substring(2, 15) + Math.random().toString(36).substring(2, 15);
-}
-
-window.addEventListener("load", init, false);
-
-</script>
-
-<textarea rows="4" cols="50" type="text" id="field1">
+<html>
+    <meta charset="utf-8" />
+    <title>snekboxweb</title>
+    <body>
+        <form action="/result" method="POST">
+            <p>Code:<br><textarea name="Code" rows="20" cols="80">
 def sum(a,b):
     return a+b
-print( sum(1,2) )
-</textarea>
-<br>
-<button onclick="sendFromInput()">Send</button>
-<button onclick="exit()">disconnect from websocket</button>
-<div id="output"></div>
+print( sum(1,2) )</textarea></p>
+            <p><input type="submit" value="Run"></p>
+        </form>
+    </body>
+</html>
