+++ templates/result.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html>
+    <meta charset="utf-8" />
+    <title>snekboxweb</title>
+    <body>
+        <p>Code Evaluated:<br><pre>{{ code }}</pre></p>
+        <p>Results:<br><pre>{{ result }}</pre></p>
+    </body>
+</html>
